[MOVE/#220] InternInfoModel 패키지 이동
