+++ domain/src/main/java/com/terning/domain/entity/intern/InternInfoModel.kt
@@ -1,4 +1,4 @@
-package com.terning.domain.entity.response
+package com.terning.domain.entity.intern
 
 data class InternInfoModel(
     val dDay: String,
