+++ domain/src/main/java/com/terning/domain/repository/InternRepository.kt
@@ -1,6 +1,6 @@
 package com.terning.domain.repository
 
-import com.terning.domain.entity.response.InternInfoModel
+import com.terning.domain.entity.intern.InternInfoModel
 
 interface InternRepository {
     suspend fun getInternInfo(id: Long): Result<InternInfoModel>
