+++ feature/src/main/java/com/terning/feature/intern/model/InternUiState.kt
@@ -3,7 +3,7 @@ package com.terning.feature.intern.model
 import androidx.compose.ui.graphics.Color
 import com.terning.core.designsystem.theme.CalRed
 import com.terning.core.state.UiState
-import com.terning.domain.entity.response.InternInfoModel
+import com.terning.domain.entity.intern.InternInfoModel
 
 data class InternUiState(
     val loadState: UiState<InternInfoModel> = UiState.Loading,
